Do not return InvalidLength error if readMsgLength() fails
As the underlying caller might be a network socket reader, a possible error could be a i/o timeout
which the underlying caller might handle accordingly. If we, instead respond with our own error, this might
break things.

Additionally, there was a bug in the ReadAll() method. We provided the unlimited reader instead of the limited reader.